+++ js/functions.js
@@ -1,10 +1,10 @@
+// Функции из этого файла не используются в проекте, но пока лежат тут для тренировки :)
+
 //Функция для проверки длины строки
-const checksLengthString = (string, maxLength) => {
-  const result = (string.length <= maxLength);
-  return result;
-};
+const checksLengthString = (string, maxLength) => string.length <= maxLength;
 
 const lengthString = checksLengthString('в этой строке 25 символов', 26);
+
 console.log(lengthString);
 
 //Функция для проверки, является ли строка палиндромом
@@ -14,10 +14,10 @@ const checkPalindrome = (string) => {
   for (let i = stringNormalize.length - 1; i >= 0; i--) {
     stringReverse += stringNormalize.at(i);
   }
-  const result = (stringReverse === stringNormalize);
-  return result;
+  return stringReverse === stringNormalize;
 };
 const palindrome = checkPalindrome('Лёша на полке клопа нашёл ');
+
 console.log(palindrome);
 
 //Функция для извлечения чисел
@@ -28,17 +28,13 @@ const extractNumbers = (string) => {
 
   for (let i = 0; i < string.length; i++) {
     const number = parseInt(string.at(i), 10);
-    if (Number.isNaN(number)) {
-      stringNumbers += '';
-    } else {
+    if (!Number.isNaN(number)) {
       stringNumbers += number;
     }
   }
-  if (parseInt(stringNumbers, 10)) {
-    return Number(stringNumbers);
-  } else {
-    return NaN;
-  }
+  return parseInt(stringNumbers, 10);
 };
-const numbers = extractNumbers(-1.24);
+const numbers = extractNumbers('f');
+const numbers1 = extractNumbers('-1asdasd.24');
 console.log(numbers);
+console.log(numbers1);
